+++ components/theme/index.tsx
@@ -1,8 +1,8 @@
 import type { CSSInterpolation } from '@ant-design/cssinjs';
 import { Theme, useCacheToken, useStyleRegister } from '@ant-design/cssinjs';
 import React from 'react';
-import version from '../../version';
-import type { DeepPartial } from '../type';
+import version from '../version';
+import type { DeepPartial } from '../_util/type';
 import type {
   AliasToken,
   DerivativeToken,
