+++ components/theme/themes/dark/colorAlgorithm.ts
