+++ components/theme/themes/default/colorAlgorithm.ts
