+++ components/theme/themes/default/index.ts
