+++ components/theme/themes/default/palettes.ts
