+++ components/theme/util/genComponentStyleHook.ts
@@ -2,7 +2,7 @@
 import type { CSSInterpolation } from '@ant-design/cssinjs';
 import { useStyleRegister } from '@ant-design/cssinjs';
 import { useContext } from 'react';
-import { ConfigContext } from '../../../config-provider/context';
+import { ConfigContext } from '../../config-provider/context';
 import type { UseComponentStyleResult } from '../index';
 import { mergeToken, statisticToken, useToken } from '../index';
 import type { GlobalToken, OverrideToken } from '../interface';
